feat: finishing challenger 01

+++ bootcamp2024/node/01-desafio-task-nodejs/src/routes.js
@@ -22,6 +22,11 @@ export const routes = [
     path: buildRoutePath("/tasks"),
     handler: (req, res) => {
       const { title, description } = req.body;
+
+      if (!title) return res.writeHead(400).end("Title is required");
+      if (!description)
+        return res.writeHead(400).end("Description is required");
+
       database.insert("tasks", {
         id: randomUUID(),
         title,
@@ -39,6 +44,13 @@ export const routes = [
     handler: (req, res) => {
       const { id } = req.params;
       const { title, description } = req.body;
+
+      if (!title) return res.writeHead(400).end("Title is required");
+      if (!description)
+        return res.writeHead(400).end("Description is required");
+      if (!database.validateId("tasks", id))
+        return res.writeHead(400).end("Task not found");
+
       database.update("tasks", id, {
         title,
         description,
@@ -52,6 +64,11 @@ export const routes = [
     path: buildRoutePath("/tasks/:id"),
     handler: (req, res) => {
       const { id } = req.params;
+
+      if (!database.validateId("tasks", id)) {
+        return res.writeHead(400).end("Task not found");
+      }
+
       database.delete("tasks", id);
       return res.writeHead(204).end();
     },
@@ -61,6 +78,11 @@ export const routes = [
     path: buildRoutePath("/tasks/:id/complete"),
     handler: (req, res) => {
       const { id } = req.params;
+
+      if (!database.validateId("tasks", id)) {
+        return res.writeHead(400).end("Task not found");
+      }
+
       database.update("tasks", id, { completed_at: new Date() });
       return res.writeHead(200).end();
     },
